backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0

openai>=1.55.0

langchain==0.2.11
langchain-openai==0.1.21
langchain-community==0.2.10
langchain-text-splitters==0.2.0

langgraph==0.2.34

faiss-cpu==1.13.1

pydantic==2.8.2
python-dotenv==1.0.1

numpy>=1.26.0

tiktoken==0.7.0

pytest==8.2.2
pytest-asyncio==0.23.0
httpx==0.27.0